chore(build): check source with Sonarcloud
As in sda-dropwizard-commons, the Sonar check will run with the definition from the master branch and is triggered by Java CI.
It loads the test reports from the artifacts published by the Java CI workflow.
This is necessary because we don't want to (and can't) execute workflows that use the code that can be updated in the PR with secrets.