fix(deps): bump spring boot and spring cloud version up
bump spring boot to 3.5.0
bump spring cloud to 2025.0.0

BREAKING CHANGE: it can happen that a misalignment occurs between the version of the junit-platform-launcher provided by gradle and the version of the junit-jupiter-engine provided in the spring dependencies. In this case, gradle is not able to run the tests and throws an error. It's necessary to add testRuntimeOnly 'org.junit.platform:junit-platform-launcher' as a dependency in your build.gradle file then. That takes the version for the unit-platform-launcher from the dependency management.